requirements.txt
fastapi==0.115.12
pydantic==2.11.2
pymongo==4.11.3
python-dotenv==1.1.0
uvicorn==0.34.0
python-multipart==0.0.20
pydantic==2.11.2
pipreqs==0.5.0